Create horizen directory and move related code there
