+++ poseidon2-starky/benches/poseidon2_starky.rs
@@ -2,9 +2,9 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use plonky2::field::types::Sample;
 use plonky2::plonk::config::{GenericConfig, PoseidonGoldilocksConfig};
 use plonky2::util::timing::TimingTree;
-use poseidon2_starky::columns::STATE_SIZE;
-use poseidon2_starky::generation::{generate_poseidon2_trace, Row};
-use poseidon2_starky::stark::{trace_to_poly_values, Poseidon2Stark};
+use poseidon2_starky::horizen::columns::STATE_SIZE;
+use poseidon2_starky::horizen::generation::{generate_poseidon2_trace, Row};
+use poseidon2_starky::horizen::stark::{trace_to_poly_values, Poseidon2Stark};
 use starky::config::StarkConfig;
 use starky::prover::prove;
 
@@ -13,7 +13,7 @@ type C = PoseidonGoldilocksConfig;
 type F = <C as GenericConfig<D>>::F;
 type S = Poseidon2Stark<F, D>;
 
-fn bench_poseidon2_starky(c: &mut Criterion) {
+fn bench_horizen_poseidon2_starky(c: &mut Criterion) {
     let mut config = StarkConfig::standard_fast_config();
     config.fri_config.cap_height = 0;
     config.fri_config.rate_bits = 3; // to meet the constraint degree bound
@@ -32,7 +32,7 @@ fn bench_poseidon2_starky(c: &mut Criterion) {
     let trace_poly_values = trace_to_poly_values(trace);
 
     let mut timing = TimingTree::default();
-    c.bench_function("poseidon2_starky", |b| {
+    c.bench_function("horizen_poseidon2_starky", |b| {
         b.iter_batched(
             || trace_poly_values.clone(),
             |trace_poly_values| {
@@ -46,6 +46,6 @@ fn bench_poseidon2_starky(c: &mut Criterion) {
 criterion_group! {
     name = benches;
     config = Criterion::default().measurement_time(std::time::Duration::from_secs(20)).sample_size(50);
-    targets = bench_poseidon2_starky
+    targets = bench_horizen_poseidon2_starky
 }
 criterion_main!(benches);
