+++ poseidon2-starky/src/lib.rs
@@ -8,6 +8,4 @@
 #![deny(clippy::cargo)]
 #![allow(clippy::missing_panics_doc)]
 
-pub mod columns;
-pub mod generation;
-pub mod stark;
+pub mod horizen;
